services/recurring-task-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
python-dateutil>=2.9.0
